fix static IP configuration for erouter0:0 interface for /32 Subnet RIP ( OFW-6025 F3896SI-2787 )
[Issue]
erouter0:0 interface is not having the static IP configured via config file

[Description]
Issue is seen because restart rip interface is calling with a false value for single static, which is not needed and it removes static ip from erouter0:0 interface.

[Solution]
Remove that portion of code executing in single static case and make it available in multi-static case only.

[Testing]
1.Upgrade the image with new change.
2.Configure the /32 Subnet RIP via bootfile.
3.Reboot
4.Once the device comes up observe the erouter0:0 interface ip

Signed-off-by: Anagha prakash <apadincharachalukulangara.ext@libertyglobal.com>