fix IPv6 LanMode value after CBN Mv1 db migration from legacy to OFW ( OFW-6100 CH7465SI-1162 )
[Issue]
During migration of IPv6LANMode value from legacy to OFW, there is an error happened with the value. In legacy software
stateless value is represented with value 0. But in OFW its corresponding value is 2. The initial patch given for IPv6LANMode
migration, it didnt mapped legacy stateless to OFW stateless(0 -> 2). instead it migrated the IPv6LANMode value as 0 in OFW also.
Later we provided with proper patch for migration, but device already migrated to OFW still holds value as 0 for stateless.

[Solution]
modems which shows as "empty" value, it'll be configured as "stateless" by setting the value as 2.

[Testing Performed]
1.Perform db migration from legacy to OFW 2.0 image
2.Make sure gpv of Device.X_LGI-COM_Gateway.IPv6LANMode gives empty value.
3.Flash new image and confirm value is showing as stateless.